patternson_runner/obspat/requirements.txt
stix2
stix2-patterns
pyyaml
simple_settings==0.13
alignment==1.0.10
numpy
pebble==4.3.10
click==7.0
